+++ Source/Components/Touchpoints/Types/Carousel/MLBusinessTouchpointsCarouselContainerItemView.swift
@@ -8,8 +8,9 @@
 import Foundation
 import MLUI
 
-class MLBusinessTouchpointsCarouselContainerItemView: UIView {
-
+final class MLBusinessTouchpointsCarouselContainerItemView: UIView {
+    fileprivate enum Configuration {}
+    
     private let containerView: UIView = {
         let containerView = UIView(frame: .zero)
         containerView.translatesAutoresizingMaskIntoConstraints = false
@@ -107,6 +108,7 @@ class MLBusinessTouchpointsCarouselContainerItemView: UIView {
     }()
 
     private var logoImageViewTopConstraint: NSLayoutConstraint?
+    private var logoImageViewbackgroundConstraint: NSLayoutConstraint?
     private var discountValueVerticalStackViewTopConstraint: NSLayoutConstraint?
     private var brandNameLabelTopConstraint: NSLayoutConstraint?
     var imageProvider: MLBusinessImageProvider = MLBusinessURLImageProvider()
@@ -124,67 +126,59 @@ class MLBusinessTouchpointsCarouselContainerItemView: UIView {
 
     private func setup() {
         addSubview(containerView)
-
         containerView.addSubview(logoImageView)
         containerView.addSubview(overlayLogoView)
         containerView.addSubview(discountValueVerticalStackView)
         containerView.addSubview(brandNameLabel)
         containerView.addSubview(subtitleLabel)
-
         addSubview(pillView)
     }
 
     private func setupConstraints() {
+
+        logoImageViewTopConstraint = logoImageView.topAnchor.constraint(equalTo: containerView.topAnchor, constant: Configuration.Layout.logoImageViewTopConstraint)
+        logoImageViewTopConstraint?.isActive = true
+        
+        discountValueVerticalStackViewTopConstraint = discountValueVerticalStackView.topAnchor.constraint(equalTo: logoImageView.bottomAnchor, constant: Configuration.Layout.discountValueVerticalStackViewTopConstraint)
+        discountValueVerticalStackViewTopConstraint?.isActive = true
+
+        brandNameLabelTopConstraint = brandNameLabel.topAnchor.constraint(equalTo: discountValueVerticalStackView.bottomAnchor, constant: Configuration.Layout.brandNameLabelTopConstraint)
+        brandNameLabelTopConstraint?.isActive = true
+        
         NSLayoutConstraint.activate([
             containerView.topAnchor.constraint(equalTo: topAnchor),
             containerView.leftAnchor.constraint(equalTo: leftAnchor),
             containerView.rightAnchor.constraint(equalTo: rightAnchor),
             containerView.bottomAnchor.constraint(equalTo: bottomAnchor),
-        ])
-
-        logoImageViewTopConstraint = logoImageView.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 16.0)
-        logoImageViewTopConstraint?.isActive = true
-
-        NSLayoutConstraint.activate([
-            logoImageView.heightAnchor.constraint(equalToConstant: 72),
+            
+            logoImageView.heightAnchor.constraint(equalToConstant: Configuration.Layout.ImageHeight),
             logoImageView.widthAnchor.constraint(equalTo: logoImageView.heightAnchor),
             logoImageView.centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
-        ])
-
-        NSLayoutConstraint.activate([
-            overlayLogoView.heightAnchor.constraint(equalToConstant: 72),
+            
+            overlayLogoView.heightAnchor.constraint(equalToConstant: Configuration.Layout.ImageHeight),
             overlayLogoView.widthAnchor.constraint(equalTo: overlayLogoView.heightAnchor),
             overlayLogoView.topAnchor.constraint(equalTo: logoImageView.topAnchor),
             overlayLogoView.centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
-        ])
-
-        discountValueVerticalStackViewTopConstraint = discountValueVerticalStackView.topAnchor.constraint(equalTo: logoImageView.bottomAnchor, constant: 12.0)
-        discountValueVerticalStackViewTopConstraint?.isActive = true
-
-        NSLayoutConstraint.activate([
-            discountValueVerticalStackView.leftAnchor.constraint(greaterThanOrEqualTo: containerView.leftAnchor, constant: 8.0),
-            discountValueVerticalStackView.rightAnchor.constraint(lessThanOrEqualTo: containerView.rightAnchor, constant: -8.0),
+            
+            overlayLogoView.heightAnchor.constraint(equalToConstant: Configuration.Layout.ImageHeight),
+            overlayLogoView.widthAnchor.constraint(equalTo: overlayLogoView.heightAnchor),
+            overlayLogoView.topAnchor.constraint(equalTo: logoImageView.topAnchor),
+            overlayLogoView.centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
+            
+            discountValueVerticalStackView.leftAnchor.constraint(greaterThanOrEqualTo: containerView.leftAnchor, constant: Configuration.Layout.leftAnchor),
+            discountValueVerticalStackView.rightAnchor.constraint(lessThanOrEqualTo: containerView.rightAnchor, constant: Configuration.Layout.rightAnchor),
             discountValueVerticalStackView.centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
-        ])
-
-        brandNameLabelTopConstraint = brandNameLabel.topAnchor.constraint(equalTo: discountValueVerticalStackView.bottomAnchor, constant: 8)
-        brandNameLabelTopConstraint?.isActive = true
-
-        NSLayoutConstraint.activate([
-            brandNameLabel.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: 8),
-            brandNameLabel.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: -8),
-        ])
-
-        NSLayoutConstraint.activate([
-            subtitleLabel.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: 8),
-            subtitleLabel.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: -8),
-            subtitleLabel.topAnchor.constraint(equalTo: brandNameLabel.bottomAnchor, constant: 2),
-            subtitleLabel.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: -12),
-        ])
-
-        NSLayoutConstraint.activate([
+            
+            brandNameLabel.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: Configuration.Layout.leftAnchor),
+            brandNameLabel.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: Configuration.Layout.rightAnchor),
+            
+            subtitleLabel.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: Configuration.Layout.leftAnchor),
+            subtitleLabel.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: Configuration.Layout.rightAnchor),
+            subtitleLabel.topAnchor.constraint(equalTo: brandNameLabel.bottomAnchor, constant: Configuration.Layout.subtitleLabelTop),
+            subtitleLabel.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: Configuration.Layout.subtitleLabelBottom),
+            
             pillView.centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
-            pillView.bottomAnchor.constraint(equalTo: logoImageView.bottomAnchor, constant: 8),
+            pillView.bottomAnchor.constraint(equalTo: logoImageView.bottomAnchor, constant: Configuration.Layout.pillViewBottom),
         ])
     }
 
@@ -250,6 +244,13 @@ class MLBusinessTouchpointsCarouselContainerItemView: UIView {
             discountValueVerticalStackViewTopConstraint?.constant = 12
             logoImageViewTopConstraint?.constant = 16
         }
+        
+        if item.type?.uppercased() == Configuration.cardType.full {
+            setTypeFull()
+            overlayLogoView.isHidden = true
+        } else {
+            overlayLogoView.isHidden = false
+        }
 
         applyFormats(to: item)
         
@@ -259,6 +260,18 @@ class MLBusinessTouchpointsCarouselContainerItemView: UIView {
     public func clear() {
         logoImageView.image = nil
         pillView.icon = nil
+    }
+    
+    private func setTypeFull() {
+        brandNameLabelTopConstraint?.isActive = false
+        logoImageView.layer.cornerRadius = 6.0
+        logoImageView.clipsToBounds = true
+        NSLayoutConstraint.activate([
+            logoImageView.topAnchor.constraint(equalTo: topAnchor),
+            logoImageView.leadingAnchor.constraint(equalTo: leadingAnchor),
+            logoImageView.trailingAnchor.constraint(equalTo: trailingAnchor),
+            logoImageView.bottomAnchor.constraint(equalTo: bottomAnchor)
+        ])
     }
     
     private func applyFormats(to item: MLBusinessCarouselItemModel) {
@@ -294,5 +307,25 @@ class MLBusinessTouchpointsCarouselContainerItemView: UIView {
         } else {
             label.font = MLStyleSheetManager.styleSheet.regularSystemFont(ofSize: size)
         }
+    }
+}
+
+extension MLBusinessTouchpointsCarouselContainerItemView.Configuration {
+    enum cardType {
+        static let full: String = "FULL"
+    }
+    
+    enum Layout {
+        static let leftAnchor: CGFloat = 8
+        static let rightAnchor: CGFloat = -8
+        static let ImageHeight: CGFloat = 72
+        static let subtitleLabelBottom:  CGFloat = -12
+        static let subtitleLabelTop:  CGFloat = 2
+        static let pillViewBottom:  CGFloat = 8
+        static let logoImageViewTopConstraint: CGFloat = 16.0
+        static let discountValueVerticalStackViewTopConstraint: CGFloat = 12
+        static let brandNameLabelTopConstraint: CGFloat = 8
+
+        
     }
 }