+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.35.0
 🚀 1.35.0 🚀
 - Fix accessibility in carousel touchpoints
+- adding a full type in Touchpoints carousel
 
 # v1.34.1
 🚀 1.34.1 🚀