+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Sin Publicar
+# v1.35.0
+🚀 1.35.0 🚀
 - Fix accessibility in carousel touchpoints
 - adding a full type in Touchpoints carousel